brute_rpn: Add support for "nonexistent candidates".
The implementation in polytope_compositor.cc is a bit ugly and could probably
be made quicker. I'll do that if required; but nonexistent candidates at least
works properly now, so ISDA should test for situations where a candidate
enters or leaves the Smith set, not just failures among candidates that stay
in the Smith set both before and after.

+++ src/singlewinner/brute_force/general_rpn/composition/equivalences.h
@@ -156,7 +156,7 @@ class fixed_cand_equivalences {
 struct election_scenario_pair {
 	std::list<ballot_group> election;
 	copeland_scenario scenario;
-	int from_perspective_of;
+	ssize_t from_perspective_of;
 };
 
 std::map<int, fixed_cand_equivalences> get_cand_equivalences(
